refactor make_model_devi
